Workerized fetching events highlights
