+++ applications/vessel-history/package.json
@@ -69,7 +69,8 @@
     "workbox-range-requests": "^5.1.3",
     "workbox-routing": "^5.1.3",
     "workbox-strategies": "^5.1.3",
-    "workbox-streams": "^5.1.3"
+    "workbox-streams": "^5.1.3",
+    "workerize-loader": "^1.3.0"
   },
   "devDependencies": {
     "@globalfishingwatch/linting": "^1.6.1",
