+++ applications/vessel-history/src/features/vessels/activity/vessels-activity.selectors.test.ts
@@ -8,7 +8,7 @@ import {
 jest.mock('features/profile/filters/filters.slice')
 
 describe('selectFilteredEvents', () => {
-  const cases: [string, string, string, RenderedEvent[][], Filters, number][] = [
+  const cases: [string, string, string, RenderedEvent[], Filters, number][] = [
     [
       'encounters',
       '2019-01-01',
